pyproject.toml
[project]
name = "task1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.0",
    "langchain>=1.2.6",
    "langchain-groq>=1.1.1",
    "langgraph>=1.0.6",
    "mcp>=1.25.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
